refactor: validation 과정 GoalStatus value로 옮김
